Allows injection of DeliveryFailed handler
Maintains the previous behavior as the default, however it is now
possible to inject into the Kafka::AsyncProducer a custom handler for
Kafka::DeliveryFailed errors. This allows the program that uses
ruby-kafka to determine a more appropriate way to handle delivery
failures if needed.

Without this ability, it is possible to end up in a situation where the
producer's buffer fills up with messages that (due to a bug,
misconfiguration, etc) can *never* be delivered to a broker, and
ultimately a DOS of the AsyncProducer, because callers will receive
buffer overflow errors until the process is restarted.

Now it is possible to use a custom handler to deal with the failed
messages individually and (potentially) clear out the producer's buffer
to allow the application to continue producing messages.