HACK - change hardcoded isolation level for consumers
FC-682

I really wanted to make this configurable on a consumer by consumer
basis, but there is a great deal of spaghetti code that needs to be
unwound in order to be able to pass the isolation level down into each
consumer. Because we need to allow reading uncommitted messages as an
emergency fix to get some of our stuck consumers running again, we're
just going to flip-flop the hardcoded value here for now.